Add functionality to help run tilesets headlessly
Added options to remove the loading of components that aren't useful on
the server:
 * `skipCamera` to not load camera objects
 * `skipGraphics` to not load animations, sprites, and other graphics

Added an option to get a different type of resource loader other than
Resource: `resourceBuilder`. This can allow for loading of files and
data through something like `node:fs` or really anything other than
`Resource` because it has a dependency on XMLHttpRequest and isn't
available outside of the browser.